+++ azure-kusto-data/azure/kusto/data/helpers.py
@@ -1,7 +1,7 @@
-import sys
-from typing import TYPE_CHECKING, Union, Callable, Dict, Optional
+import json
 from functools import lru_cache
-import copy
+from pathlib import Path
+from typing import TYPE_CHECKING, Union, Callable, Dict, Optional
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -9,6 +9,12 @@ if TYPE_CHECKING:
 
 # Alias for dataframe_from_result_table converter type
 Converter = Dict[str, Union[str, Callable[[str, "pd.DataFrame"], "pd.Series"]]]
+
+
+def load_bundled_json(file_name: str) -> Dict:
+    filename = Path(__file__).absolute().parent.joinpath(file_name)
+    with filename.open("r", encoding="utf-8") as data:
+        return json.load(data)
 
 
 @lru_cache(maxsize=1, typed=False)