+++ azure-kusto-ingest/tests/test_e2e_ingest.py
@@ -529,14 +529,19 @@ class TestE2E:
             "xtextWithNulls",
             "xdynamicWithNulls",
         ]
-        rows = [
-            [0, "00000000-0000-0000-0001-020304050607", 0.0, 0.0, 0, 0, 0, 0, 0, 0, 0, 0, "2014-01-01T01:01:01Z", "Zero", "Zero", "0", "00:00:00", None, ""]
-        ]
+
+        guid = uuid.uuid4()
+
+        dynamic_value = ["me@dummy.com", "you@dummy.com", "them@dummy.com"]
+        rows = [[0, str(guid), 0.0, 0.0, 0, 0, 0, 0, 0, 0, 0, 0, "2014-01-01T01:01:01Z", "Zero", "Zero", "0", "00:00:00", None, dynamic_value]]
         df = DataFrame(data=rows, columns=fields)
-        ingestion_properties = IngestionProperties(database=self.test_db, table=self.test_table, flush_immediately=True, data_format=DataFormat.CSV)
+        ingestion_properties = IngestionProperties(database=self.test_db, table=self.test_table, flush_immediately=True)
         self.ingest_client.ingest_from_dataframe(df, ingestion_properties)
 
         await self.assert_rows_added(1, timeout=120)
+
+        a = self.client.execute(self.test_db, f"{self.test_table} | where rowguid == '{guid}'")
+        assert a.primary_results[0].rows[0]["xdynamicWithNulls"] == dynamic_value
 
     @pytest.mark.asyncio
     async def test_streaming_ingest_from_blob(self, is_managed_streaming):