+++ CHANGELOG.md
@@ -5,6 +5,50 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.3] - 2025-05-04
+
+### Fixed
+
+- Fixed Ingest from dataframe
+- Dependency issues
+- Typing issues
+
+## [5.0.2] - 2025-05-30
+
+### Fixed
+
+- Fixed bug in `ingest_from_dataframe` with some ingestion properties combinations.
+
+## [5.0.1] - 2025-03-12
+
+### Fixed
+
+- Dependency issues
+
+## [5.0.0] - 2025-03-10
+
+### Changed
+- [BREAKING] Dropped support for python version 3.7, as it is on EOL for over year.
+- [BREAKING] Aligned the Connection String Builder keywords with the rest of the SDKs.
+This means that some keywords were removed, and they will no longer be parsed as part of the Connection String.  
+Building the Connection String using the builder method will still work as expected.  
+The following keywords have been removed:
+    - `msi_auth` / `msi_authentication`
+    - `msi_params` / `msi_type`
+    - `interactive_login`
+    - `az_cli`
+- [BREAKING] `ingest_from_dataframe` - Added `data_format` parameter. It can be None (default), 'json' or 'csv'.
+  Based on how panda's csv serialization works, dynamic data will not be serialized correctly.
+  By default, the data will be serialized as json to avoid this issue.
+  However, that may break if a CSV mapping is used.
+  Therefore, when passing the None value, the data will be json by default, or csv if a csv mapping is used.
+  Also, it is possible to pass 'csv' or 'json' to force the serialization type.
+
+## [4.6.3] - 2025-01-08
+
+### Fixed
+- Explicitly export members in `__init__.py` via `__all__`
+
 ## [4.6.2] - 2025-01-07
 
 ### Fixed